feat: Update Modulator to operate in time domain
Refactor the ModulatorParams class to include a new attribute "temporal" of type TemporalParams. This allows the ModulatorParams to operate in the time domain. The TemporalParams class defines parameters related to the temporal characteristics of the modulator, such as the time axis and center frequency.

Update the ModulationType enum to handle different modulation types and include a setup_target method for initialization.

Remove commented out code for the ArbitraryModulator class.